# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Senior Voice API
  description: 适老化语音问答 App 后端接口（FastAPI 契约）
  version: 0.1.0
servers:
  - url: http://192.168.1.100:8000/api/v1
    description: 本地/局域网开发（与 config/api.json 保持一致）
...
paths:
...
  /auth/sms/send:
    post:
      tags: [auth]
      summary: 发送短信验证码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendSmsRequest'
      responses:
        '200':
          description: 已发送
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendSmsResponse'
        '429':
          $ref: '#/components/responses/Error'
  /auth/register:
    post:
      tags: [auth]
      summary: 手机号注册（验证码 + 密码，可选昵称）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功并返回登录态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /auth/login/sms:
    post:
      tags: [auth]
      summary: 验证码登录（可同时设置密码）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /auth/login/password:
    post:
      tags: [auth]
      summary: 密码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Error'
  /auth/password:
    post:
      tags: [auth]
      summary: 修改或设置密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '401':
          $ref: '#/components/responses/Error'
...
  /auth/token/refresh:
    post:
      tags: [auth]
      summary: 刷新 access_token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 新的 token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Error'
...
  /archives/ocr:
    post:
      tags: [archives]
      summary: 病历图片 OCR（就诊单或体检单）
      description: |
        大模型先判断 document_type：visit 写入 medical_archives，exam 写入 health_reports。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, source]
              properties:
                file:
                  type: string
                  format: binary
                source:
                  type: string
                  enum: [camera, album]
      responses:
        '200':
          description: OCR 结构化结果（已入库）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrResult'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
  /archives:
    get:
      tags: [archives]
      summary: 就诊单列表（时间轴另一半）
      description: |
        只返回就诊单。前端会与 GET /health-reports 合并后按日期倒序画时间轴。
        卡片使用 diagnosis / medicine / visit_date / visit_no。
        ORDER BY visit_date DESC, created_at DESC。
        q 前端展示页不传，可暂缓。
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 前端展示页传 100
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveListResponse'
    post:
      tags: [archives]
      summary: 保存档案
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArchiveRequest'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveRecord'
  /archives/{archive_id}:
    get:
      tags: [archives]
      summary: 就诊单详情
      description: 完整报告 Tab 使用 raw_ocr_text；校验所属用户。
      parameters:
        - $ref: '#/components/parameters/ArchiveId'
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveRecord'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [archives]
      summary: 更新档案
      parameters:
        - $ref: '#/components/parameters/ArchiveId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArchiveRequest'
      responses:
        '200':
          description: 更新后的档案
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArchiveRecord'
    delete:
      tags: [archives]
      summary: 删除档案
      parameters:
        - $ref: '#/components/parameters/ArchiveId'
      responses:
        '200':
          description: 已删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /archives/{archive_id}/share:
    post:
      tags: [archives]
      summary: 推送给家属
      parameters:
        - $ref: '#/components/parameters/ArchiveId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareArchiveRequest'
      responses:
        '200':
          description: 推送结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareArchiveResponse'
  /archives/{archive_id}/export:
    get:
      tags: [archives]
      summary: 导出 PDF 下载链接
      parameters:
        - $ref: '#/components/parameters/ArchiveId'
      responses:
        '200':
          description: 下载信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPdfResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ArchiveId:
      name: archive_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
    Lang:
      type: string
      enum: [zh, en]
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    SendSmsRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          example: '13800138000'
        purpose:
          type: string
          enum: [login, register, reset_password]
          default: login
    RegisterRequest:
      type: object
      required: [phone, code, password]
      properties:
        phone:
          type: string
        code:
          type: string
        password:
          type: string
          minLength: 6
        display_name:
          type: string
          maxLength: 64
        preferred_lang:
          type: string
          enum: [zh, en]
          default: zh
    SendSmsResponse:
      type: object
      required: [ok, expire_in]
      properties:
        ok:
          type: boolean
        expire_in:
          type: integer
          description: 验证码有效秒数
    SmsLoginRequest:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
        password:
          type: string
          description: 可选，首次设置密码
    PasswordLoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    ChangePasswordRequest:
      type: object
      required: [new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
    UserProfile:
      type: object
      required: [id, phone, preferred_lang, created_at]
      properties:
        id:
          type: string
        phone:
          type: string
        display_name:
          type: string
          nullable: true
        preferred_lang:
          $ref: '#/components/schemas/Lang'
        created_at:
          type: string
          format: date-time
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/TokenPair'
        - type: object
          required: [user]
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
...
    RiskLevel:
      type: string
      enum: [low, medium, high]
      nullable: true
...
    OcrFinding:
      type: object
      required: [title, suggestion, sort_order]
      properties:
        title:
          type: string
        suggestion:
          type: string
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        sort_order:
          type: integer
    OcrResult:
      type: object
      required: [document_type, diagnosis, medicine, visit_date, visit_no, raw_ocr_text]
      properties:
        document_type:
          type: string
          enum: [visit, exam]
          description: visit=就诊单（medical_archives）；exam=体检单（health_reports）
        id:
          type: string
          nullable: true
          description: 已落库记录 id（就诊单 archive id / 体检单 report id）
        diagnosis:
          type: string
        medicine:
          type: string
        visit_date:
          type: string
          format: date
        visit_no:
          type: string
          description: 就诊号/门诊号；体检单时等同 voucher_no
        raw_ocr_text:
          type: string
        patient_name:
          type: string
          nullable: true
        org_name:
          type: string
          nullable: true
        voucher_no:
          type: string
          nullable: true
        report_type:
          type: string
          nullable: true
        findings:
          type: array
          items:
            $ref: '#/components/schemas/OcrFinding'
    ArchiveRecord:
      type: object
      required: [id, diagnosis, medicine, visit_date, visit_no, created_at, updated_at]
      properties:
        id:
          type: string
        diagnosis:
          type: string
        medicine:
          type: string
        visit_date:
          type: string
          format: date
        visit_no:
          type: string
          description: 就诊号/门诊号，用户维度业务唯一标识
        raw_ocr_text:
          type: string
          nullable: true
          description: 详情「完整报告」使用；没有则 null
        image_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ArchiveListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArchiveRecord'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    CreateArchiveRequest:
      type: object
      required: [diagnosis, medicine, visit_date, visit_no]
      properties:
        diagnosis:
          type: string
        medicine:
          type: string
        visit_date:
          type: string
          format: date
        visit_no:
          type: string
        raw_ocr_text:
          type: string
        image_url:
          type: string
    UpdateArchiveRequest:
      type: object
      properties:
        diagnosis:
          type: string
        medicine:
          type: string
        visit_date:
          type: string
          format: date
        visit_no:
          type: string
    ShareArchiveRequest:
      type: object
      required: [contact_ids]
      properties:
        contact_ids:
          type: array
          items:
            type: string
        message:
          type: string
    ShareArchiveResponse:
      type: object
      required: [ok, shared_count]
      properties:
        ok:
          type: boolean
        shared_count:
          type: integer
    ExportPdfResponse:
      type: object
      required: [download_url, expires_in]
      properties:
        download_url:
          type: string
          format: uri
        expires_in:
          type: integer
...
security:
  - bearerAuth: []